refactored question components
