+++ client/src/pages/NewQuestionPage.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import NewQuestionForm from '../components/QuestionComponents/NewQuestionForm'
+
+const NewQuestionPage = () => {
+  return(
+    <div>
+      <NewQuestionForm />
+    </div>
+  )
+}
+
+export default NewQuestionPage;
